backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
motor==3.3.2
python-multipart==0.0.9
openai==1.14.2
pdf2image==1.17.0
PyMuPDF==1.24.1
python-dotenv==1.0.1
